Add types to package.json
