+++ packages/third-parties/package.json
@@ -8,6 +8,7 @@
   "exports": {
     "./google": "./dist/google/index.js"
   },
+  "types": "dist/types/google.d.ts",
   "files": [
     "dist"
   ],
