+++ packages/third-parties/src/types/google.ts
@@ -1,3 +1,18 @@
+declare global {
+  interface Window {
+    dataLayer?: Object[]
+    [key: string]: any
+  }
+}
+
+export type GTMParams = {
+  gtmId: string
+  dataLayer?: string[]
+  dataLayerName?: string
+  auth?: string
+  preview?: string
+}
+
 export type GoogleMapsEmbed = {
   height?: number
   width?: number
